Create README.md
During this project, I created and executed a basic application for managing caches. The application includes the implementation of two cache management strategies: FIFO (First-In, First-Out) and LFU (Least Frequently Used). The purpose is to effectively handle a cache with a predetermined capacity. The program receives page requests from the user and verifies if each page is already present in the cache. If a page is not found, the code selects a page to remove from the cache based on the chosen policy: the first-in page for FIFO or the least frequently used page for LFU. At the conclusion of each policy's execution, the contents of the cache are displayed.